% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookout-package.R
\docType{package}
\name{lookout-package}
\alias{lookout-package}
\title{lookout: Leave One Out Kernel Density Estimates for Outlier Detection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Outlier detection using leave-one-out kernel density estimates and extreme value theory. The bandwidth for the kernel density estimates is computed from the edge lengths of the Euclidean minimum spanning tree of the data. Using the peak-over-threshold method, a generalized Pareto distribution is fitted to the negative log of the leave-one-out kernel density estimates to identify outliers. See Kandanaarachchi and Hyndman (2022) \doi{10.1080/10618600.2021.2000425} and Hyndman, Kandanaarachchi and Turner (2026) \doi{10.48550/arXiv.2603.22636}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sevvandi.github.io/lookout/}
  \item \url{https://github.com/sevvandi/lookout}
  \item Report bugs at \url{https://github.com/sevvandi/lookout/issues}
}

}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

Authors:
\itemize{
  \item Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})
  \item Rob Hyndman \email{rob.hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
}

Other contributors:
\itemize{
  \item Chris Fraley \email{fraley@u.washington.edu} [contributor]
}

}
\keyword{internal}
