% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loookoutliers.R
\name{lookoutliers}
\alias{lookoutliers}
\title{Identifies outliers using the algorithm lookout.}
\usage{
lookoutliers(X, alpha = 0.05, unitize = TRUE, bw = NULL)
}
\arguments{
\item{X}{The input data in a dataframe, matrix or tibble format.}

\item{alpha}{The level of significance. Default is \code{0.0.05}.}

\item{unitize}{An option to normalize the data. Default is \code{TRUE}, which normalizes each column to \code{[0,1]}.}

\item{bw}{Bandwidth parameter. Default is \code{NULL} as the bandwidth is found using Persistent Homology.}
}
\value{
A list with the following components:
\item{\code{outliers}}{The set of outliers.}
\item{\code{outlier_probability}}{The GPD probability of the data.}
\item{\code{bandwidth}}{The bandwdith selected using persistent homology. }
\item{\code{kde}}{The kernel density estimate values.}
\item{\code{lookde}}{The leave-one-out kde values.}
\item{\code{gpd}}{The fitted GPD parameters.}
}
\description{
This function identifies outliers using the algorithm lookout, an outlier detection method that uses leave-one-out kernel density estimates and generalized Pareto distributions to find outliers.
}
\examples{
set.seed(1)
x1 <- cbind.data.frame(rnorm(500), rnorm(500))
x2 <- cbind.data.frame(rnorm(5, mean = 10, sd = 0.2), rnorm(5, mean = 10, sd = 0.2))
colnames(x1) <- colnames(x2) <- c("x", "y")
X <- rbind.data.frame(x1, x2)
plot(X, pch = 19)
out <- lookoutliers(X)
}
