% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_persistence.R
\name{persisting_outliers_over_alpha}
\alias{persisting_outliers_over_alpha}
\title{Computes outlier persistence for a range of significance values.}
\usage{
persisting_outliers_over_alpha(
  X,
  alpha_min = 0.01,
  alpha_max = 0.1,
  step = 0.01,
  st_qq = 0.9,
  en_sc = sqrt(5),
  unitize = TRUE,
  num_steps = 20
)
}
\arguments{
\item{X}{The input data in a dataframe, matrix or tibble format.}

\item{alpha_min}{The smallest level of significance, set to \code{0.01}.}

\item{alpha_max}{The maximum level of significance, set to \code{0.1}.}

\item{step}{The step size, set to \code{0.01}.}

\item{st_qq}{The starting quantile for death radii sequence. This will be used to compute the starting bandwidth value.}

\item{en_sc}{The scaling for the maximum death radii. This will be used to compute the biggest bandwidth value.}

\item{unitize}{An option to normalize the data. Default is \code{TRUE}, which normalizes each column to \code{[0,1]}.}

\item{num_steps}{The length of the bandwidth sequence.}
}
\value{
A list with the following components:
\item{\code{out}}{A 3D array of \code{N x num_steps x num_alpha} where \code{N} denotes the number of observations, num_steps denote the length of the bandwidth sequence and num_alpha denotes the number of significance levels. This is a binary array and the entries are set to 1 if that observation is an outlier for that particular bandwidth and significance level.}
\item{\code{bw}}{The set of bandwidth values.}
\item{\code{gpdparas}}{The GPD parameters used. }
\item{\code{lookoutbw}}{The bandwidth chosen by the algorithm \code{lookout} using persistent homology.}
}
\description{
This function computes outlier persistence for a range of significance values, using the algorithm lookout, an outlier detection method that uses leave-one-out kernel density estimates and generalized Pareto distributions to find outliers.
}
\examples{
set.seed(1)
x1 <- cbind.data.frame(rnorm(500), rnorm(500))
x2 <- cbind.data.frame(rnorm(5, mean=10, sd=0.2), rnorm(5, mean=10, sd=0.2))
colnames(x1) <- colnames(x2) <- c("x", "y")
X <- rbind.data.frame(x1, x2)
plot(X, pch=19)
outnew <- persisting_outliers_over_alpha(X, unitize=FALSE)
}
