% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_outlier_barcode}
\alias{plot_outlier_barcode}
\title{Plots outlier persistence for a given significance level.}
\usage{
plot_outlier_barcode(outnew, alpha = 0.05)
}
\arguments{
\item{outnew}{The output of the function \code{persisting_outliers_over_alpha}.}

\item{alpha}{The level of significance, set to \code{0.05}.}
}
\value{
A ggplot object.
}
\description{
This function plots outlier persistence for a given significance level. using the algorithm lookout, an outlier detection method that uses leave-one-out kernel density estimates and generalized Pareto distributions to find outliers.
}
\examples{
set.seed(1)
x1 <- cbind.data.frame(rnorm(500), rnorm(500))
x2 <- cbind.data.frame(rnorm(5, mean = 10, sd = 0.2), rnorm(5, mean = 10, sd = 0.2))
colnames(x1) <- colnames(x2) <- c("x", "y")
X <- rbind.data.frame(x1, x2)
plot(X, pch = 19)
outnew <- persisting_outliers_over_alpha(X, unitize = FALSE)
plot_outlier_barcode(outnew, alpha = 0.05)
}
