% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_outlier_barcode_over_alpha}
\alias{plot_outlier_barcode_over_alpha}
\title{Plots outlier persistence for a range of significance levels.}
\usage{
plot_outlier_barcode_over_alpha(outnew)
}
\arguments{
\item{outnew}{The output of the function \code{persisting_outliers_over_alpha}.}
}
\value{
A ggplot object.
}
\description{
This function plots outlier persistence a range of significance levels.
}
\examples{
set.seed(1)
x1 <- cbind.data.frame(rnorm(500), rnorm(500))
x2 <- cbind.data.frame(rnorm(5, mean=10, sd=0.2), rnorm(5, mean=10, sd=0.2))
colnames(x1) <- colnames(x2) <- c("x", "y")
X <- rbind.data.frame(x1, x2)
plot(X, pch=19)
outnew <- persisting_outliers_over_alpha(X, unitize=FALSE)
plot_outlier_barcode_over_alpha(outnew)

}
