% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{find_tda_bw}
\alias{find_tda_bw}
\title{Identifies bandwidth for outlier detection.}
\usage{
find_tda_bw(X, fast = TRUE, bw_para = 0.95)
}
\arguments{
\item{X}{The input data in a dataframe, matrix or tibble format.}

\item{fast}{If set to \code{TRUE}, makes the computation faster by sub-setting
the data for the bandwidth calculation.}

\item{bw_para}{Parameter for bandwidth calculation. Default is \code{0.95}.
If set to 1, then the bandwidth corresponds to the maximum Rips death radi
difference. If set to 0.95, then the bandwidth corresponds to the 95th
quantile of Rips death radi. Other probabilities can be used.}
}
\value{
The bandwidth
}
\description{
This function identifies the bandwidth that is used in the kernel density
estimate computation. The function uses topological data analysis (TDA)
to find the badnwidth.
}
\examples{
X <- rbind(
  data.frame(x = rnorm(500),
             y = rnorm(500)),
  data.frame(x = rnorm(5, mean = 10, sd = 0.2),
             y = rnorm(5, mean = 10, sd = 0.2))
)
find_tda_bw(X, fast = TRUE)

}
