% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookout-package.R
\docType{package}
\name{lookout-package}
\alias{lookout-package}
\title{lookout: Leave One Out Kernel Density Estimates for Outlier Detection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Outlier detection using leave-one-out kernel density estimates and extreme value theory. The bandwidth for kernel density estimates is computed using persistent homology, a technique in topological data analysis. Using peak-over-threshold method, a generalized Pareto distribution is fitted to the log of leave-one-out kde values to identify outliers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sevvandi.github.io/lookout/}
}

}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

Authors:
\itemize{
  \item Rob Hyndman \email{rob.hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
}

}
\keyword{internal}
