% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutliers.R
\name{lookout}
\alias{lookout}
\title{Identifies outliers using the algorithm lookout.}
\usage{
lookout(
  X,
  alpha = 0.05,
  unitize = FALSE,
  normalize = TRUE,
  bw = NULL,
  gpd = NULL,
  fast = TRUE,
  bw_para = 0.95,
  shape_zero = TRUE
)
}
\arguments{
\item{X}{The numerical input data in a data.frame, matrix or tibble format.}

\item{alpha}{The level of significance. Default is \code{0.05}.}

\item{unitize}{If \code{TRUE}, the data is standardized so that
each column is in the range \code{[0,1]}. Default is \code{FALSE}.}

\item{normalize}{If \code{TRUE} (default), each column of the data is
transformed to be closer to a standard normal distribution, and the data
is rotated so that the columns are pairwise uncorrelated. This is done
robustly using \code{\link{weird}{mvscale}}.}

\item{bw}{Bandwidth parameter. If \code{NULL} (default), the bandwidth is
found using Persistent Homology.}

\item{gpd}{Generalized Pareto distribution parameters. If `NULL` (the
default), these are estimated from the data.}

\item{fast}{If \code{TRUE} (default), makes the computation faster by
sub-setting the data for the bandwidth calculation.}

\item{bw_para}{Parameter for bandwidth calculation. Default is \code{0.95}.
If set to 1, then the bandwidth corresponds to the maximum Rips death radii
difference. If set to 0.95, then the bandwidth corresponds to the 95th
quantile of Rips death radii. Other probabilities can be used.}

\item{shape_zero}{If \code{TRUE} (default), the shape parameter in the GPD is
set to zero resulting in a Gumbel distribution. Default is \code{TRUE}.}
}
\value{
A list with the following components:
\item{\code{outliers}}{The set of outliers.}
\item{\code{outlier_probability}}{The GPD probability of the data.}
\item{\code{outlier_scores}}{The outlier scores of the data.}
\item{\code{bandwidth}}{The bandwdith selected using persistent homology. }
\item{\code{kde}}{The kernel density estimate values.}
\item{\code{lookde}}{The leave-one-out kde values.}
\item{\code{gpd}}{The fitted GPD parameters.}
}
\description{
This function identifies outliers using the algorithm lookout, an outlier
detection method that uses leave-one-out kernel density estimates and
generalized Pareto distributions to find outliers.
}
\examples{
X <- rbind(
  data.frame(
    x = rnorm(500),
    y = rnorm(500)
  ),
  data.frame(
    x = rnorm(5, mean = 10, sd = 0.2),
    y = rnorm(5, mean = 10, sd = 0.2)
  )
)
lo <- lookout(X)
lo
autoplot(lo)
}
